Correct get shard status logic
